feat(a380-sd): visual rework of APU page
